requirements.txt
asgiref==3.2.3
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.0.5
django-ckeditor==6.0.0
django-cors-headers==3.2.1
django-heroku==0.3.1
django-js-asset==1.2.2
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysqlclient==1.4.6
numpy==1.19.4
pandas==1.1.5
Pillow==8.0.1
psycopg2==2.8.5
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.10
whitenoise==5.2.0
